% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_afl_ground.R
\name{annotate_afl_oval}
\alias{annotate_afl_oval}
\title{Annotate AFL Oval}
\usage{
annotate_afl_oval(
  dim,
  origin = list(x = 0, y = 0),
  aspect_ratio = 1,
  limits = TRUE,
  colour = "dimgray",
  fill = NA,
  size = 1,
  ...
)
}
\arguments{
\item{dim}{TODO}

\item{origin}{TODO}

\item{aspect_ratio}{TODO}

\item{limits}{TODO}

\item{colour}{TODO}

\item{fill}{TODO}

\item{size}{TODO}

\item{...}{TODO}
}
\value{
list of ggplot geoms to be added to a ggplot plot
}
\description{
Description to be added
}
\examples{
library(ggplot2)

# Create the MCG
g <- ggplot() +
     annotate_afl_oval(dim = ground_mcg, fill = "green", alpha = 0.2)

# Add blank theme
g + theme_ground()

# Create the Gabba
ggplot() +
annotate_afl_oval(dim = ground_gabba, fill = "green", alpha = 0.2) +
theme_ground()

# Custom ground
dim <- list(length = 175, width = 130, origin_x = 0, origin_y = 0)

ggplot() +
annotate_afl_oval(dim = dim, fill = "green", alpha = 0.2) +
theme_ground()


}
