% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AusSportGrounds-package.R
\docType{package}
\name{AusSportGrounds-package}
\alias{AusSportGrounds}
\alias{AusSportGrounds-package}
\title{AusSportGrounds: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SportStatisticsRSweet/AusSportGrounds}
  \item Report bugs at \url{https://github.com/SportStatisticsRSweet/AusSportGrounds/issues}
}

}
\author{
\strong{Maintainer}: James Day \email{jamesthomasday@gmail.com} (\href{https://orcid.org/0000-0003-3133-943X}{ORCID})

Authors:
\itemize{
  \item Alice Sweeting
}

}
\keyword{internal}
