% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ground.R
\name{theme_ground}
\alias{theme_ground}
\title{Removes background and axes details from a ggplot plot.}
\usage{
theme_ground()
}
\value{
list of ggplot themes to be added to a ggplot plot
}
\description{
Functionally very similar to \code{ggplot2::theme_void} and
adapted directly from \code{ggsoccer::theme_pitch}.
}
\examples{
library(ggplot2)

data <- data.frame(x = c(90, 85, 82, 78, 83),
                         y = c(43, 40, 52, 56, 44))

p <- ggplot(data, aes(x = x, y = y)) +
  geom_point()

# Remove all elements from plot
p + theme_ground()



}
