% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_dimensions.R
\docType{data}
\name{ground_mcg}
\alias{ground_mcg}
\alias{ground_gabba}
\title{Ground dimensions}
\format{
An object of class \code{list} of length 4.

An object of class \code{list} of length 4.
}
\usage{
ground_mcg

ground_gabba
}
\description{
The dimensions of the ground to build with the \code{annotate_} set of functions.

We provide some common ground dimensions but custom dimensions can be used.
}
\details{
A "ground dimensions" is simply a list of dimensions that define a
coordinate system. The required dimensions are:

\itemize{
\item{"length"}{The length of the longest part of the ground (x axis)}
\item{"width"}{The width of the ground from side to the other (y axis)}
\item{"origin_x"}{The minimum x coordinate of the ground}
\item{"origin_y"}{The minimum y coordinate of the ground}
}

The following ground dimensions are provided
\itemize{
\item{"ground_mcg"}{For the MCG}
}
}
\keyword{datasets}
